requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-mysql==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
appnope==0.1.3
argcomplete==2.0.0
asttokens==2.0.5
attrs==20.3.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
cachelib==0.7.0
cattrs==1.10.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-DebugToolbar==0.13.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipykernel==6.13.1
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jsonschema==4.6.0
jupyter-client==7.3.3
jupyter-core==4.10.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prison==0.2.1
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
psycopg2-binary==2.9.3
psycopg2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.1.0
requests==2.27.1
rfc3986==1.5.0
rich==12.4.4
setproctitle==1.2.3
six==1.16.0
snakeviz==2.1.1
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
